Moved Scroll box to _components
